Make changes necessary to support multiple driver environments
This commit encompases a lot of work and changes that all result in
making the av-api capable of understanding "where" it is deployed and
translating that into how it will communicate with different drivers
given the deployment environment.

This work also, coincidentally, breaks almost all of the dependencies
that the av-api has on common, with the notable exceptions of log,
events, and nerr in a couple of places.